PerAccountNotificationSettingsGroup [nfc]: Extract out to this new file
