+++ src/settings/NotificationsScreen.js
@@ -8,10 +8,7 @@ import invariant from 'invariant';
 import type { RouteProp } from '../react-navigation';
 import type { AppNavigationProp } from '../nav/AppNavigator';
 import { useGlobalSelector, useSelector } from '../react-redux';
-import { getAuth, getSettings } from '../selectors';
-import SwitchRow from '../common/SwitchRow';
 import Screen from '../common/Screen';
-import * as api from '../api';
 import ServerPushSetupBanner from '../common/ServerPushSetupBanner';
 import NavRow from '../common/NavRow';
 import { IconAlertTriangle } from '../common/Icons';
@@ -19,15 +16,14 @@ import type { LocalizableText } from '../types';
 import { TranslationContext } from '../boot/TranslationProvider';
 import { kWarningColor } from '../styles/constants';
 import { getIdentities, getIdentity, getIsActiveAccount } from '../account/accountsSelectors';
-import { getRealm, getRealmName } from '../directSelectors';
-import ZulipText from '../common/ZulipText';
-import SettingsGroup from './SettingsGroup';
+import { getRealm } from '../directSelectors';
 import { openSystemNotificationSettings } from '../utils/openLink';
 import {
   useNotificationReportsByIdentityKey,
   NotificationProblem,
 } from './NotifTroubleshootingScreen';
 import { keyOfIdentity } from '../account/accountMisc';
+import PerAccountNotificationSettingsGroup from './PerAccountNotificationSettingsGroup';
 
 type Props = $ReadOnly<{|
   navigation: AppNavigationProp<'notifications'>,
@@ -56,7 +52,6 @@ export default function NotificationsScreen(props: Props): Node {
 
   const _ = useContext(TranslationContext);
 
-  const auth = useSelector(getAuth);
   const identity = useSelector(getIdentity);
   const notificationReportsByIdentityKey = useNotificationReportsByIdentityKey();
   const notificationReport = notificationReportsByIdentityKey.get(keyOfIdentity(identity));
@@ -67,11 +62,7 @@ export default function NotificationsScreen(props: Props): Node {
   const otherAccounts = useGlobalSelector(state =>
     getIdentities(state).filter(identity_ => !getIsActiveAccount(state, identity_)),
   );
-  const realmName = useSelector(getRealmName);
   const pushNotificationsEnabled = useSelector(state => getRealm(state).pushNotificationsEnabled);
-  const offlineNotification = useSelector(state => getSettings(state).offlineNotification);
-  const onlineNotification = useSelector(state => getSettings(state).onlineNotification);
-  const streamNotification = useSelector(state => getSettings(state).streamNotification);
 
   const systemSettingsWarnings = notificationReport.problems
     .map(systemSettingsWarning)
@@ -100,37 +91,6 @@ export default function NotificationsScreen(props: Props): Node {
 
     openSystemNotificationSettings();
   }, [systemSettingsWarnings, _]);
-
-  const handleTroubleshootingPress = useCallback(() => {
-    navigation.push('notif-troubleshooting');
-  }, [navigation]);
-
-  // TODO(#3999): It'd be good to show "working on it" UI feedback while a
-  //   request is pending, after the user touches a switch.
-
-  const handleOfflineNotificationChange = useCallback(() => {
-    api.toggleMobilePushSettings({
-      auth,
-      opp: 'offline_notification_change',
-      value: !offlineNotification,
-    });
-  }, [offlineNotification, auth]);
-
-  const handleOnlineNotificationChange = useCallback(() => {
-    api.toggleMobilePushSettings({
-      auth,
-      opp: 'online_notification_change',
-      value: !onlineNotification,
-    });
-  }, [onlineNotification, auth]);
-
-  const handleStreamNotificationChange = useCallback(() => {
-    api.toggleMobilePushSettings({
-      auth,
-      opp: 'stream_notification_change',
-      value: !streamNotification,
-    });
-  }, [streamNotification, auth]);
 
   const handleOtherAccountsPress = useCallback(() => {
     navigation.push('account-pick');
@@ -166,44 +126,7 @@ export default function NotificationsScreen(props: Props): Node {
       {!notificationReport.problems.includes(NotificationProblem.SystemSettingsDisabled) && (
         <>
           {pushNotificationsEnabled && (
-            <SettingsGroup
-              title={{
-                text: 'Notification settings for this account ({email} in {realmName}):',
-                values: {
-                  email: <ZulipText style={{ fontWeight: 'bold' }} text={identity.email} />,
-                  realmName: <ZulipText style={{ fontWeight: 'bold' }} text={realmName} />,
-                },
-              }}
-            >
-              <SwitchRow
-                label="Notifications when offline"
-                value={offlineNotification}
-                onValueChange={handleOfflineNotificationChange}
-              />
-              <SwitchRow
-                label="Notifications when online"
-                value={onlineNotification}
-                onValueChange={handleOnlineNotificationChange}
-              />
-              <SwitchRow
-                label="Stream notifications"
-                value={streamNotification}
-                onValueChange={handleStreamNotificationChange}
-              />
-              <NavRow
-                {...(() =>
-                  notificationReport.problems.length > 0 && {
-                    leftElement: {
-                      type: 'icon',
-                      Component: IconAlertTriangle,
-                      color: kWarningColor,
-                    },
-                    subtitle: 'Notifications for this account may not arrive.',
-                  })()}
-                title="Troubleshooting"
-                onPress={handleTroubleshootingPress}
-              />
-            </SettingsGroup>
+            <PerAccountNotificationSettingsGroup navigation={navigation} />
           )}
           {otherAccounts.length > 0 && (
             <NavRow
